decorator refactor
