--- ngoto/core/constants.py
@@ -1,2 +1,0 @@
-plugin_path: str = 'ngoto/plugins/'
-command_path: str = "ngoto/cogs"
