+++ ngoto/core/util/node.py
@@ -66,7 +66,7 @@ class Node:
 
     def get_children(self) -> list:
         return self.children
-    
+
     def get_child_from_name(self, name: str):
         for child in self.children:
             if child.name == name:
@@ -75,7 +75,7 @@ class Node:
 
     def get_child(self, index: int):  # returns Node obj
         return self.children[index]
-    
+
     def has_child(self, name: str) -> bool:
         for child in self.children:
             if child.name == name or child.name.lower() == name.lower():
